refactored all chat service demos
